+++ weblablib.py
@@ -912,17 +912,17 @@ class WebLab(object):
         if tasks:
             return tasks[0]
 
-    def join_tasks(self, func_or_name, timeout=None, stopping=True):
+    def join_tasks(self, func_or_name, timeout=None, stop=False):
         """
         Stop (optionally) and join all the tasks with a given name.
 
         :param func_or_name: you can either provide the task function or its name (string)
         :param timeout: seconds to wait. By default wait forever.
-        :param stopping: call ``stop()`` to each task before call ``join()``. 
+        :param stop: call ``stop()`` to each task before call ``join()``. 
         """
         tasks = self.get_running_tasks(func_or_name)
 
-        if stopping:
+        if stop:
             for task in tasks:
                 task.stop()